uid-range: use size_t for array size

+++ src/shared/uid-range.h
@@ -8,8 +8,8 @@ typedef struct UidRange {
         uid_t start, nr;
 } UidRange;
 
-int uid_range_add(UidRange **p, unsigned *n, uid_t start, uid_t nr);
-int uid_range_add_str(UidRange **p, unsigned *n, const char *s);
+int uid_range_add(UidRange **p, size_t *n, uid_t start, uid_t nr);
+int uid_range_add_str(UidRange **p, size_t *n, const char *s);
 
-int uid_range_next_lower(const UidRange *p, unsigned n, uid_t *uid);
-bool uid_range_contains(const UidRange *p, unsigned n, uid_t uid);
+int uid_range_next_lower(const UidRange *p, size_t n, uid_t *uid);
+bool uid_range_contains(const UidRange *p, size_t n, uid_t uid);
